informs/requirements.txt
django~=5.1
gunicorn
django_bootstrap_icons
django-tz-detect
crispy-bootstrap5
django-crispy-forms
folium
geopy
icecream
